// pkgs/frontend/src/db/schema.ts
export interface Database {
// ...
  public: {
    Tables: {
      amm_pairs: {
        Row: {
          id: number
          lptokencode: string
          token1currency: string
          token1issuer: string | null
          token2currency: string
          token2issuer: string | null
        }
        Insert: {
          id?: number
          lptokencode: string
          token1currency: string
          token1issuer?: string | null
          token2currency: string
          token2issuer?: string | null
        }
        Update: {
          id?: number
          lptokencode?: string
          token1currency?: string
          token1issuer?: string | null
          token2currency?: string
          token2issuer?: string | null
        }
        Relationships: []
      }
      carbon_credit_tokens: {
        Row: {
          currency: string | null
          framework: string | null
          id: number
          issuer: string | null
        }
        Insert: {
          currency?: string | null
          framework?: string | null
          id?: number
          issuer?: string | null
        }
        Update: {
          currency?: string | null
          framework?: string | null
          id?: number
          issuer?: string | null
        }
        Relationships: []
      }
      credit_frameworks: {
        Row: {
          framework: string
          id: number
        }
        Insert: {
          framework: string
          id?: number
        }
        Update: {
          framework?: string
          id?: number
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}